fix: precision, gating, severity, and introspection batch
- Hessian/gradient accumulation, the GN solve, and the degeneracy eigen-
  analysis now run in double. Summing 10k+ float outer products loses
  significant digits exactly where the degeneracy gate needs to
  discriminate near-singular blocks; per-point math stays float.

- Intensity range correction (and reflectivity extraction) now only run
  when the photometric term is active (photometricWeight > 0). The
  corrected intensities propagate into the published deskewed/keyframe
  clouds, so downstream consumers were silently receiving modified
  sensor data even with the feature disabled -- plus a per-point pow()
  for nothing.

- 'Bad time sync' and 'low number of points' were logged FATAL but both
  degrade gracefully (rigid transform / skipped scan); now throttled
  WARNs with accurate messages.

- updateKeyframes decision cascade documented with its truth table (the
  rotation trigger only applies in the num_nearby <= 1 carve-out).

- savePCD talks through the node logger instead of std::cout.

- dlio::declare_param gained a descriptor overload; key tunables
  (photometric, degeneracy, regularization, pruning, dashboard) now have
  descriptions surfaced by 'ros2 param describe'.

Verified on Jazzy: clean build, 4/4 gtests, composed launch loads both
components, descriptions visible via ros2 param describe.

+++ include/nano_gicp/nano_gicp.h
@@ -87,7 +87,7 @@ public:
 protected:
   virtual void computeTransformation(PointCloudSource& output, const Eigen::Matrix4f& guess) override;
 
-  void linearize(const Eigen::Isometry3f& trans, Eigen::Matrix<float, 6, 6>* H, Eigen::Matrix<float, 6, 1>* b);
+  void linearize(const Eigen::Isometry3f& trans, Eigen::Matrix<double, 6, 6>* H, Eigen::Matrix<double, 6, 1>* b);
   void update_correspondences(const Eigen::Isometry3f& trans);
 
   template<typename PointT>
